Use owned callback for send_identity
This allows user data associated with the callback to be kept
referenced until the callback is replaced.

It seems that there are other unsolved problems with this part
of the code currently, in that the callback is called whenever the
'send' button is pressed even when no actual query is active, which
currently leads to crashes.