configure.ac: Fix dependencies checking
The pkg-config command is discovered through a separate macro,
PKG_PROG_PKG_CONFIG that takes care of identifying the presence (and version)
of pkg-config itself. This macro is called through AC_REQUIRE
so that is expanded before PKG_CHECK_MODULES.

If we have the first call to PKG_CHECK_MODULES inside a bash conditional block,
the expansion of PKG_PROG_PKG_CONFIG will also be conditional,
so the previous code will fail to work.