Fix 'make dist'
This is hairy; conditionally compiling .vala files breaks 'make dist'
because every possible generated .c file is required. I've worked
around the problem by putting all the RPC server code into one file
with a big #if, so that all of the .vala files are always compiled.
It's a bit ugly though.