https://bugs.launchpad.net/moonshot-ui/+bug/1447008 Moonshot UI window becomes too tall and is not-resizeable
Used ScrolledWindow to add a scrollbar to the services list, and ensured that
the list would still grow vertically if the user expands the window.
Made other changes to ensure that things that should *not* grow vertically on
resise, don't.

Also changed the way the details pane switches between showing No_Identity and
other cards. This eliminated some redundant code that had allowed bugs to
creep in during this effort.

Fixed formatting of the services table when it has a small number of entries.

Removed some unnecessary logging.