Enable interactive CLI UI only when in the foreground
The interactive CLI UI implies that the end user might be
prompted to introduce information (passwords,
accept trust anchor...). If the application using Moonshot
is not executed in the foreground, it won't have access to
the terminal STDOUT and STDIN and hence, will block.

This change makes sure that CLI UI is only executed when the
application has control over STDIN and STDOUT.