Avoid storing all the IdCards after each action (ADD, DEL, UPDATE).
All the IdCards were stored after every single action involving
one card. This was probably a reminiscence of the FlatFile format,
where all cards need to be stored becase it is a sequential file.

The keyring allows updating/removing/adding individual IdCards without
touching the rest of the Keyring. Hence, a HashMap has been added to
the Keyring subclasses so they track the association between IdCards
provided to the UI and the internal reference for the keyring: item_id
for the GKR and Item object for libsecret